refactor: AddExpenseCategorySection 리팩토링
