+++ src/app/home/setting/expenseCategory/_components/AddExpenseCategorySection.type.tsx
@@ -1,0 +1,9 @@
+import { ExpenseCategoryAction } from "./Main";
+
+export interface AddExpenseCategorySectionProps {
+    expenseCategoriesDispatch: React.Dispatch<ExpenseCategoryAction>;
+}
+
+export interface UseAddExpenseCategorySectionProps {
+    expenseCategoriesDispatch: React.Dispatch<ExpenseCategoryAction>;
+}
